Crud working good

+++ contactsapi/src/main/java/tech/cristianciobanu/contactsapi/contact/ContactService.java
@@ -1,22 +1,26 @@
 package tech.cristianciobanu.contactsapi.contact;
 
+import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
+import tech.cristianciobanu.contactsapi.auth.exception.NotAuthorizedException;
 import tech.cristianciobanu.contactsapi.skill.Skill;
 import tech.cristianciobanu.contactsapi.skill.SkillDto;
+import tech.cristianciobanu.contactsapi.skill.SkillRepository;
+import tech.cristianciobanu.contactsapi.user.JpaUserDetailsService;
+import tech.cristianciobanu.contactsapi.user.UserRepository;
 
 import java.util.*;
 
 @Service
+@RequiredArgsConstructor
 public class ContactService {
-
     private final ContactRepository contactRepository;
+    private final SkillRepository skillRepository;
+    private final UserRepository userRepository;
+    private final JpaUserDetailsService userDetailsService;
     private final ContactMapper contactMapper;
-
-
-    public ContactService(ContactRepository contactRepository, ContactMapper contactMapper) {
-        this.contactRepository = contactRepository;
-        this.contactMapper = contactMapper;
-    }
 
     public List<ContactDto> findAll() {
         List<ContactDto> contactList = new ArrayList<>();
@@ -37,18 +41,32 @@ public class ContactService {
     }
 
     public ContactDto create(ContactDto contact) {
+        var loggedUsername = SecurityContextHolder.getContext().getAuthentication().getName();
+        contact.setCreatedBy(loggedUsername);
+        contact.setSkills(this.setContactSkills(contact));
+
         Contact createdContact = contactRepository.save(contactMapper.contactDtoToContact(contact));
         return contactMapper.contactToContactDto(createdContact);
     }
 
     public Optional<ContactDto> update(UUID id, ContactDto updatedContact) {
+        var loggedUsername = SecurityContextHolder.getContext().getAuthentication().getName();
         return contactRepository.findById(id)
                 .map(foundContact -> {
+                    if (!foundContact.getCreatedBy().equals(loggedUsername)){
+                        throw new NotAuthorizedException("User is not authorized to modify this contact");
+                    }
                     Contact newContact = updateContactWith(foundContact, updatedContact);
                     return contactMapper.contactToContactDto(contactRepository.save(newContact));});
     }
 
     public void delete(UUID id) {
+        var loggedUsername = SecurityContextHolder.getContext().getAuthentication().getName();
+        var foundContact = contactRepository.findById(id)
+                        .orElseThrow(() -> new RuntimeException("Contact not found"));
+        if (!foundContact.getCreatedBy().equals(loggedUsername)){
+            throw new NotAuthorizedException("User is not authorized to modify this contact");
+        }
         contactRepository.deleteById(id);
     }
 
@@ -61,7 +79,23 @@ public class ContactService {
                 newContact.getAddress(),
                 newContact.getEmail(),
                 newContact.getPhoneNumber(),
-                new HashSet<>(newContact.getSkills())
+                setContactSkills(newContact),
+                oldContact.getCreatedBy()
         );
+    }
+
+    private Set<Skill> setContactSkills(ContactDto contact){
+        var loggedUsername = SecurityContextHolder.getContext().getAuthentication().getName();
+        HashSet<Skill> skills = new HashSet<>();
+        contact.getSkills().forEach(skill -> {
+            Skill foundSkill = this.skillRepository.findSkillByNameContainingIgnoreCaseAndLevel(skill.getName(), skill.getLevel());
+            if (foundSkill != null) {
+                skills.add(foundSkill);
+            } else {
+                skill.setCreatedBy(loggedUsername);
+                skills.add(this.skillRepository.save(skill));
+            }
+        });
+        return skills;
     }
 }
